refactor: changed namespace for ProcessingRecoverableError
