+++ triples-generator/src/main/scala/io/renku/triplesgenerator/events/categories/triplesgenerated/projectinfo/ProjectInfoFinder.scala
@@ -27,7 +27,7 @@ import io.renku.control.Throttler
 import io.renku.graph.model.entities.Project.{GitLabProjectInfo, ProjectMember}
 import io.renku.graph.model.projects
 import io.renku.http.client.AccessToken
-import io.renku.triplesgenerator.events.categories.Errors.ProcessingRecoverableError
+import io.renku.triplesgenerator.events.categories.ProcessingRecoverableError
 import org.typelevel.log4cats.Logger
 
 private[triplesgenerated] trait ProjectInfoFinder[F[_]] {
@@ -53,9 +53,9 @@ private[triplesgenerated] class ProjectInfoFinderImpl[F[_]: MonadThrow: Parallel
     memberEmailFinder: MemberEmailFinder[F]
 ) extends ProjectInfoFinder[F] {
 
+  import memberEmailFinder._
   import membersFinder._
   import projectFinder._
-  import memberEmailFinder._
 
   override def findProjectInfo(
       path:                    projects.Path
