+++ triples-generator/src/main/scala/io/renku/triplesgenerator/events/categories/triplesgenerated/triplesuploading/UpdatesUploader.scala
@@ -26,7 +26,7 @@ import eu.timepit.refined.numeric.NonNegative
 import io.renku.http.client.RestClient.{MaxRetriesAfterConnectionTimeout, SleepAfterConnectionIssue}
 import io.renku.http.client.RestClientError.BadRequestException
 import io.renku.rdfstore._
-import io.renku.triplesgenerator.events.categories.Errors.ProcessingRecoverableError
+import io.renku.triplesgenerator.events.categories.ProcessingRecoverableError
 import io.renku.triplesgenerator.events.categories.triplesgenerated.RecoverableErrorsRecovery
 import io.renku.triplesgenerator.events.categories.triplesgenerated.triplesuploading.TriplesUploadResult.NonRecoverableFailure
 import org.typelevel.log4cats.Logger
