+++ triples-generator/src/test/scala/io/renku/triplesgenerator/events/categories/awaitinggeneration/triplesgeneration/renkulog/RenkuSpec.scala
@@ -24,7 +24,7 @@ import io.renku.generators.Generators.Implicits._
 import io.renku.generators.Generators._
 import io.renku.generators.jsonld.JsonLDGenerators.jsonLDEntities
 import io.renku.testtools.IOSpec
-import io.renku.triplesgenerator.events.categories.Errors.LogWorthyRecoverableError
+import io.renku.triplesgenerator.events.categories.ProcessingRecoverableError._
 import io.renku.triplesgenerator.events.categories.awaitinggeneration.triplesgeneration.renkulog.Commands.{RenkuImpl, RepositoryPath}
 import org.scalamock.scalatest.MockFactory
 import org.scalatest.matchers.should
