+++ triples-generator/src/test/scala/io/renku/triplesgenerator/events/categories/triplesgenerated/transformation/datasets/DatasetTransformerSpec.scala
@@ -27,7 +27,7 @@ import io.renku.graph.model.datasets.{InternalSameAs, ResourceId, TopmostSameAs}
 import io.renku.graph.model.testentities._
 import io.renku.graph.model.{entities, users}
 import io.renku.rdfstore.SparqlQuery
-import io.renku.triplesgenerator.events.categories.Errors.ProcessingRecoverableError
+import io.renku.triplesgenerator.events.categories.ProcessingRecoverableError
 import io.renku.triplesgenerator.events.categories.triplesgenerated.ProjectFunctions
 import io.renku.triplesgenerator.events.categories.triplesgenerated.TransformationStep.Queries
 import io.renku.triplesgenerator.events.categories.triplesgenerated.transformation.Generators.recoverableClientErrors
