+++ triples-generator/src/test/scala/io/renku/triplesgenerator/events/categories/triplesgenerated/triplesuploading/TriplesUploaderSpec.scala
@@ -34,7 +34,7 @@ import io.renku.logging.TestExecutionTimeRecorder
 import io.renku.rdfstore.{FusekiBaseUrl, SparqlQueryTimeRecorder}
 import io.renku.stubbing.ExternalServiceStubbing
 import io.renku.testtools.IOSpec
-import io.renku.triplesgenerator.events.categories.Errors.{AuthRecoverableError, LogWorthyRecoverableError}
+import io.renku.triplesgenerator.events.categories.ProcessingRecoverableError._
 import io.renku.triplesgenerator.events.categories.triplesgenerated.triplesuploading.TriplesUploadResult._
 import org.http4s.Status._
 import org.scalamock.scalatest.MockFactory
