+++ triples-generator/src/test/scala/io/renku/triplesgenerator/events/categories/triplesgenerated/triplesuploading/UpdatesUploaderSpec.scala
@@ -33,7 +33,7 @@ import io.renku.logging.TestExecutionTimeRecorder
 import io.renku.rdfstore.{FusekiBaseUrl, SparqlQuery, SparqlQueryTimeRecorder}
 import io.renku.stubbing.ExternalServiceStubbing
 import io.renku.testtools.IOSpec
-import io.renku.triplesgenerator.events.categories.Errors.{AuthRecoverableError, LogWorthyRecoverableError}
+import io.renku.triplesgenerator.events.categories.ProcessingRecoverableError._
 import io.renku.triplesgenerator.events.categories.triplesgenerated.triplesuploading.TriplesUploadResult._
 import org.http4s.Status.{Forbidden, Unauthorized}
 import org.scalatest.matchers.should
